Refactoring: add super interface for plugin features, document plugin api.
