--- api/src/main/java/org/itsallcode/whiterabbit/api/ProgressMonitor.java
@@ -1,12 +1,0 @@
-package org.itsallcode.whiterabbit.api;
-
-public interface ProgressMonitor
-{
-    boolean isCanceled();
-
-    void beginTask(String name, int totalWork);
-
-    void setTaskName(String name);
-
-    void worked(int work);
-}
