--- api/src/main/java/org/itsallcode/whiterabbit/api/ProjectReportExporter.java
@@ -1,8 +1,0 @@
-package org.itsallcode.whiterabbit.api;
-
-import org.itsallcode.whiterabbit.api.model.ProjectReport;
-
-public interface ProjectReportExporter
-{
-    void export(ProjectReport report, ProgressMonitor progressMonitor);
-}
