+++ api/src/main/java/org/itsallcode/whiterabbit/api/features/PluginFeature.java
@@ -1,0 +1,11 @@
+package org.itsallcode.whiterabbit.api.features;
+
+import org.itsallcode.whiterabbit.api.Plugin;
+
+/**
+ * Super interface for all features supported by a {@link Plugin}.
+ */
+public interface PluginFeature
+{
+
+}
