+++ api/src/main/java/org/itsallcode/whiterabbit/api/model/ActivityData.java
@@ -2,6 +2,10 @@ package org.itsallcode.whiterabbit.api.model;
 
 import java.time.Duration;
 
+/**
+ * Model for a project activity on a given {@link DayData day} including project
+ * id, duration and comment.
+ */
 public interface ActivityData
 {
     String getProjectId();
