+++ api/src/main/java/org/itsallcode/whiterabbit/api/model/DayType.java
@@ -1,5 +1,8 @@
 package org.itsallcode.whiterabbit.api.model;
 
+/**
+ * The type of a {@link DayData day}.
+ */
 public enum DayType
 {
     HOLIDAY(false), VACATION(false), FLEX_TIME(true), SICK(false), WORK(true), WEEKEND(false);
@@ -11,6 +14,9 @@ public enum DayType
         this.workDay = workDay;
     }
 
+    /**
+     * @return <code>true</code> if you need to work on a day of this type.
+     */
     public boolean isWorkDay()
     {
         return workDay;
