+++ api/src/main/java/org/itsallcode/whiterabbit/api/model/Project.java
@@ -1,5 +1,9 @@
 package org.itsallcode.whiterabbit.api.model;
 
+/**
+ * The project you work on during an {@link ProjectReportActivity activity}
+ * containing an ID, label and cost carrier id.
+ */
 public interface Project
 {
     String getProjectId();
