+++ api/src/main/java/org/itsallcode/whiterabbit/api/model/ProjectReport.java
@@ -3,6 +3,9 @@ package org.itsallcode.whiterabbit.api.model;
 import java.time.YearMonth;
 import java.util.List;
 
+/**
+ * A monthly project report.
+ */
 public interface ProjectReport
 {
     YearMonth getMonth();
