+++ api/src/main/java/org/itsallcode/whiterabbit/api/model/ProjectReportActivity.java
@@ -2,6 +2,10 @@ package org.itsallcode.whiterabbit.api.model;
 
 import java.time.Duration;
 
+/**
+ * An activity during a {@link ProjectReportDay day} where you worked for a
+ * specific time duration on a {@link Project}.
+ */
 public interface ProjectReportActivity
 {
     Project getProject();
