+++ api/src/main/java/org/itsallcode/whiterabbit/api/model/ProjectReportDay.java
@@ -3,6 +3,10 @@ package org.itsallcode.whiterabbit.api.model;
 import java.time.LocalDate;
 import java.util.List;
 
+/**
+ * The projects worked on during a single day of the {@link ProjectReport}
+ * including date and {@link ProjectReportActivity activities}.
+ */
 public interface ProjectReportDay
 {
     LocalDate getDate();
