+++ jfxui/src/main/java/org/itsallcode/whiterabbit/jfxui/ui/widget/ProgressDialog.java
@@ -2,7 +2,7 @@ package org.itsallcode.whiterabbit.jfxui.ui.widget;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.itsallcode.whiterabbit.api.ProgressMonitor;
+import org.itsallcode.whiterabbit.api.features.ProgressMonitor;
 import org.itsallcode.whiterabbit.jfxui.JavaFxUtil;
 import org.itsallcode.whiterabbit.jfxui.ui.UiResources;
