+++ logic/src/main/java/org/itsallcode/whiterabbit/logic/storage/data/JsonFileStorage.java
@@ -18,7 +18,7 @@ import javax.json.bind.Jsonb;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.itsallcode.whiterabbit.api.MonthDataStorage;
+import org.itsallcode.whiterabbit.api.features.MonthDataStorage;
 import org.itsallcode.whiterabbit.api.model.MonthData;
 import org.itsallcode.whiterabbit.logic.Config;
