+++ logic/src/test/java/org/itsallcode/whiterabbit/logic/model/MonthIndexTest.java
@@ -9,7 +9,7 @@ import java.time.LocalTime;
 import java.time.Month;
 import java.time.YearMonth;
 
-import org.itsallcode.whiterabbit.api.MonthDataStorage.ModelFactory;
+import org.itsallcode.whiterabbit.api.features.MonthDataStorage.ModelFactory;
 import org.itsallcode.whiterabbit.api.model.DayData;
 import org.itsallcode.whiterabbit.api.model.DayType;
 import org.itsallcode.whiterabbit.api.model.MonthData;
