+++ plugins/demo/src/main/java/org/itsallcode/whiterabbit/plugin/demo/DemoProjectReportExporter.java
@@ -2,8 +2,8 @@ package org.itsallcode.whiterabbit.plugin.demo;
 
 import java.time.Duration;
 
-import org.itsallcode.whiterabbit.api.ProgressMonitor;
-import org.itsallcode.whiterabbit.api.ProjectReportExporter;
+import org.itsallcode.whiterabbit.api.features.ProgressMonitor;
+import org.itsallcode.whiterabbit.api.features.ProjectReportExporter;
 import org.itsallcode.whiterabbit.api.model.ProjectReport;
 import org.itsallcode.whiterabbit.api.model.ProjectReportActivity;
 import org.itsallcode.whiterabbit.api.model.ProjectReportDay;
